+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New Features:
 
-* Tabs for systems are now sorted alphabetically by name.
+* Tabs for systems are now sorted alphabetically by name, prioritising systems that have any BGS activity first.
 * Now use scrollable tabs and a drop-down tab selector.
 
 ### Changes: